% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{inla.wkt_set_lengthunit}
\alias{inla.wkt_set_lengthunit}
\alias{inla.crs_set_lengthunit}
\title{Handling CRS/WKT}
\usage{
inla.wkt_set_lengthunit(wkt, unit, params = NULL)

inla.crs_set_lengthunit(crs, unit, params = NULL)
}
\value{
For \code{inla.wkt_set_lengthunit}, a
WKT2 string with altered length units.
Note that the length unit for the ellipsoid radius is unchanged.

For \code{inla.crs_set_lengthunit}, a \code{sp::CRS} object with
altered length units.
Note that the length unit for the ellipsoid radius is unchanged.
}
\description{
Get and set CRS object or WKT string properties.
}
